MIG delivery_carrier_label_seur
