+++ delivery_carrier_label_seur/README.rst
@@ -83,6 +83,7 @@ Contributors
 ------------
 
 * Ismael Calvo <ismael.calvo@factorlibre.com>
+* Angel Moya <angel.moya@pesol.es>
 
 Maintainer
 ----------
