+++ delivery_carrier_label_seur/__init__.py
@@ -1,2 +1,3 @@
+# -*- coding: utf-8 -*-
 from . import model
 from . import wizard
