+++ delivery_carrier_label_seur/__manifest__.py
@@ -1,13 +1,14 @@
 # -*- coding: utf-8 -*-
 # © 2015 FactorLibre - Ismael Calvo <ismael.calvo@factorlibre.com>
+# © 2017 PESOL - Angel Moya <angel.moya@pesol.es>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 {
     "name": "Seur Deliveries WebService",
     "summary": "Allows to generate shipping label for SEUR shipments.",
-    "version": "8.0.1.0.0",
+    "version": "10.0.1.0.0",
     "category": "Delivery",
     "website": "http://factorlibre.com",
-    "author": "FactorLibre, Odoo Community Association (OCA)",
+    "author": "FactorLibre, PESOL, Odoo Community Association (OCA)",
     "license": "AGPL-3",
     "application": False,
     "installable": True,
