+++ delivery_carrier_label_seur/view/delivery_view.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data>
+
+        <record id="view_delivery_carrier_form" model="ir.ui.view">
+            <field name="name">delivery_base.delivery.carrier.view_form</field>
+            <field name="model">delivery.carrier</field>
+            <field name="inherit_id" ref="base_delivery_carrier_label.view_delivery_carrier_form"/>
+            <field name="arch" type="xml">
+                <notebook position="inside">
+                    <page string="Seur" attrs="{'invisible': [('carrier_type', '!=', 'seur')]}">
+                        <group>
+                            <field name="seur_config_id" attrs="{'required': [('carrier_type', '=', 'seur')]}"/>
+                            <field name="seur_service_code" attrs="{'required': [('carrier_type', '=', 'seur')]}"/>
+                            <field name="seur_product_code" attrs="{'required': [('carrier_type', '=', 'seur')]}"/>
+                        </group>
+                    </page>
+                </notebook>
+            </field>
+        </record>
+
+    </data>
+</openerp>
