+++ delivery_carrier_label_seur/view/seur_config_view.xml
