+++ delivery_carrier_label_seur/view/stock_view.xml
