--- delivery_carrier_seur/view/delivery_view.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-
-        <record id="view_delivery_carrier_form" model="ir.ui.view">
-          <field name="name">delivery_base.delivery.carrier.view_form</field>
-          <field name="model">delivery.carrier</field>
-          <field name="inherit_id" ref="base_delivery_carrier_label.view_delivery_carrier_form" />
-          <field name="arch" type="xml">
-            <field name="type" position="after">
-              <field name="seur_config_id" attrs="{'invisible': [('type', '!=', 'seur')], 'required': [('type', '=', 'seur')]}"/>
-              <field name="seur_service_code" attrs="{'invisible': [('type', '!=', 'seur')], 'required': [('type', '=', 'seur')]}"/>
-              <field name="seur_product_code" attrs="{'invisible': [('type', '!=', 'seur')], 'required': [('type', '=', 'seur')]}"/>
-            </field>
-
-            <field name="code" position="attributes">
-                <attribute name="attrs">{'invisible': [('type', '=', 'seur')]}</attribute>
-            </field>
-          </field>
-        </record>
-
-    </data>
-</openerp>
